Rename shared-components to shared
